requirements.txt
IPython
langchain
langchain-openai
langgraph
openai
openpyxl
pandas
pydantic
python-dotenv
streamlit